fix(gui): 清理调试输出与格式警告，F04 门禁通过
- 移除装载调试 fprintf 与测试残留 DBG
- loader_verify 格式警告修复
- 门禁：clean 构建 0 错误 0 警告（上游 1 处除外）、ctest 10/10、
  Loomark.app 打包产物生成并冒烟启动正常

+++ examples/scintilla_prototype/loader_verify.cpp
@@ -121,7 +121,7 @@ int main(int argc, char* argv[])
             offset = r.nextOffset;
         }
         printf("search: hits=%llu batches=%d elapsedMs=%lld\n",
-            (unsigned long long)totalHits, batches, timer.elapsed());
+            (unsigned long long)totalHits, static_cast<int>(batches), timer.elapsed());
 
         // 取消验证：预置取消标志，必须立即返回
         std::atomic_bool cancel{true};
@@ -230,7 +230,7 @@ int main(int argc, char* argv[])
         const auto bufferLen = editor.send(SCI_GETTEXTLENGTH);
         printf("f02: diskSize=%llu bufferLen=%lld markerHits=%u\n",
             (unsigned long long)diskSize, (long long)bufferLen,
-            markerAfter.hits.size());
+            static_cast<unsigned>(markerAfter.hits.size()));
         const bool pass = snap.version > r.newVersion &&
             markerAfter.hits.size() == 1 &&
             diskSize == (unsigned long long)bufferLen;
